feat: add 1-week package

+++ FitnessCal.BLL/DTO/PaymentDTO/PaymentDtos.cs
@@ -24,7 +24,7 @@ namespace FitnessCal.BLL.DTO.PaymentDTO
         public string? Status { get; set; }
         public decimal Amount { get; set; }
         public string? PackageName { get; set; }
-        public int? DurationMonths { get; set; }
+        public decimal? DurationMonths { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? PaidAt { get; set; }
     }
@@ -39,7 +39,7 @@ namespace FitnessCal.BLL.DTO.PaymentDTO
         public string UserEmail { get; set; } = null!;
         public string UserName { get; set; } = null!;
         public string PackageName { get; set; } = null!;
-        public int DurationMonths { get; set; }
+        public decimal DurationMonths { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
     }
